Support checkpoint mechanism for verfication
We use the overflow of performance counter to track
the trace checkpoints, let's say we set trace interval
is 1000 instructions, then the overflow should happen
on each 1000 instruction.

However, in current KVM implementation, the overflow
handler is executed probably a few instructions after
the actual overflow, so there might be a tiny variance
on the checkpoint's exact instruction number.